+++ src/Product/Models/ProductAvailabilityScopeProxy.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Product\Models;
+
+use Konekt\Concord\Proxies\EnumProxy;
+
+class ProductAvailabilityScopeProxy extends EnumProxy
+{
+}